Update composer from 1.8.4 to 1.8.5 (#3269)

+++ packages/composer.rb
@@ -3,9 +3,9 @@ require 'package'
 class Composer < Package
   description 'Dependency Manager for PHP'
   homepage 'https://getcomposer.org/'
-  version '1.8.4'
-  source_url 'https://github.com/composer/composer/archive/1.8.4.tar.gz'
-  source_sha256 '288ab33c8f11f0db4b5883d4a115a8ead8ef1a74c924f3accadc61d220ca22de'
+  version '1.8.5'
+  source_url 'https://github.com/composer/composer/archive/1.8.5.tar.gz'
+  source_sha256 '923e87df4617de6cfed53c5a190623d3801cffe6178ac77c6ca559841967f714'
 
   binary_url ({
   })
@@ -16,7 +16,7 @@ class Composer < Package
   depends_on 'xdg_base'
 
   def self.preinstall
-    if Dir.exists? "#{HOME}/.config"
+    if Dir.exists?("#{HOME}/.config") && !File.symlink?("#{HOME}/.config")
       # Save any existing configuration
       system "cp -r #{HOME}/.config #{CREW_PREFIX}" unless Dir.empty? "#{HOME}/.config"
     else
